refactor(yield-yak): optimize, remove unnecessary helper
