+++ src/apps/yield-yak/avalanche/yield-yak.vault.token-fetcher.ts
@@ -1,5 +1,6 @@
 import { Inject } from '@nestjs/common';
 import { ethers, BigNumber } from 'ethers';
+import Axios from 'axios';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
@@ -13,7 +14,18 @@ import {
 } from '~position/template/app-token.template.types';
 
 import { YieldYakContractFactory, YieldYakVault } from '../contracts';
-import { YieldYakVaultTokenDefinitionsResolver } from '../helpers/yield-yak.vault.token-definitions-resolver';
+
+export type YieldYakFarmDetails = {
+  address: string;
+  deployed: number;
+  name: string;
+  depositToken: {
+    address: string;
+    symbol: string;
+    decimals: number;
+  };
+  totalDeposits: string;
+};
 
 @PositionTemplate()
 export class AvalancheYieldyakVaultTokenFetcher extends AppTokenTemplatePositionFetcher<YieldYakVault> {
@@ -23,19 +35,8 @@ export class AvalancheYieldyakVaultTokenFetcher extends AppTokenTemplatePosition
   constructor(
     @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
     @Inject(YieldYakContractFactory) private readonly contractFactory: YieldYakContractFactory,
-    @Inject(YieldYakVaultTokenDefinitionsResolver)
-    protected readonly tokenDefinitionsResolver: YieldYakVaultTokenDefinitionsResolver,
   ) {
     super(appToolkit);
-  }
-
-  private getVaultDefinitions() {
-    return this.tokenDefinitionsResolver.getVaultDefinitionsData(this.network);
-  }
-
-  private async selectVault(vaultAddress: string) {
-    const vaultDefinitions = await this.getVaultDefinitions();
-    return vaultDefinitions.find(v => v.id.toLowerCase() === vaultAddress) ?? null;
   }
 
   getContract(address: string): YieldYakVault {
@@ -43,15 +44,12 @@ export class AvalancheYieldyakVaultTokenFetcher extends AppTokenTemplatePosition
   }
 
   async getAddresses(): Promise<string[]> {
-    const vaultDefinitions = await this.getVaultDefinitions();
-    return vaultDefinitions.map(address => address.id.toLowerCase());
+    const farms = await Axios.get<YieldYakFarmDetails[]>('https://staging-api.yieldyak.com/farms').then(x => x.data);
+    return farms.map(farm => farm.address.toLowerCase());
   }
 
   async getUnderlyingTokenAddresses({ contract }: GetUnderlyingTokensParams<YieldYakVault>) {
-    const vault = await this.selectVault(contract.address.toLowerCase());
-    if (!vault) throw new Error('Cannot find specified vault');
-
-    return [vault.depositToken.id.toLowerCase()];
+    return contract.depositToken().then(addr => addr.toLowerCase());
   }
 
   async getLabel({ appToken }: GetDisplayPropsParams<YieldYakVault>): Promise<string> {
